+++ design-system/components/src/react-client/index.ts
@@ -1,1 +1,5 @@
 export { ThemeToggle } from '../react/ThemeToggle/index';
+export { ActionToggle } from '../react/ActionToggle/index';
+export { LikeButton } from '../react/LikeButton/index';
+export { Modal } from '../react/Modal/index';
+export { ConsentBanner } from '../react/ConsentBanner/index';