+++ design-system/components/src/react-server/index.ts
@@ -1,3 +1,6 @@
 export { Button } from '../react/Button/index';
 export { Card } from '../react/Card/index';
 export { Tag } from '../react/Tag/index';
+export { Breadcrumb } from '../react/Breadcrumb/index';
+export { Image } from '../react/Image/index';
+export { VerifiedBadge } from '../react/VerifiedBadge/index';